tests/test_regex_util.py
import unittest
from networkx import MultiDiGraph, MultiGraph

from project.g_util import build_two_cycle_labeled_graph, read_graph_from_file
from project.regex_util import regex_to_min_dfa, graph_to_nfa
from pyformlang.finite_automaton import Symbol, State, DeterministicFiniteAutomaton, NondeterministicFiniteAutomaton


class RegexUtilDFATest(unittest.TestCase):
    def setUp(self):
        pass

    def test_creating_min_dfa_from_regex(self):
        test_regex = "a b"
        actual = regex_to_min_dfa(test_regex)

        state_0 = State(0)
        state_1 = State(1)
        state_2 = State(2)
        symbol_a = Symbol("a")
        symbol_b = Symbol("b")

        expected = DeterministicFiniteAutomaton()
        expected.add_start_state(state_0)
        expected.add_final_state(state_2)
        expected.add_transition(state_0, symbol_a, state_1)
        expected.add_transition(state_1, symbol_b, state_2)

        assert expected.is_equivalent_to(actual)
        assert expected.minimize().is_equivalent_to(expected)

    def test_min_dfa_language(self):
        test_regex = "a b"
        dfa = regex_to_min_dfa(test_regex)
        accepted = [
            [Symbol("a"), Symbol("b")]
        ]
        not_accepted = [
            [Symbol("b"), Symbol("c")],
            [Symbol("")],
            [Symbol("a"), Symbol("c")],
            [Symbol("")],
            [Symbol("b"), Symbol("a")]
        ]
        assert all(dfa.accepts(word) for word in accepted)
        assert not all(dfa.accepts(word) for word in not_accepted)

    def test_creating_min_dfa_from_regex_with_star(self):
        test_regex = "a b c*"
        actual = regex_to_min_dfa(test_regex)

        state_0 = State(0)
        state_1 = State(1)
        state_2 = State(2)
        symbol_a = Symbol("a")
        symbol_b = Symbol("b")
        symbol_c = Symbol("c")

        expected = DeterministicFiniteAutomaton()
        expected.add_start_state(state_0)
        expected.add_final_state(state_2)
        expected.add_transition(state_0, symbol_a, state_1)
        expected.add_transition(state_1, symbol_b, state_2)
        expected.add_transition(state_2, symbol_c, state_2)

        assert expected.is_equivalent_to(actual)
        assert expected.minimize().is_equivalent_to(expected)

    def test_min_dfa_language_with_star(self):
        test_regex = "a* b* c* d*"
        dfa = regex_to_min_dfa(test_regex)
        accepted = [
            [Symbol("a")], [Symbol("b")], [Symbol("c")], [Symbol("d")],
            [Symbol("a"), Symbol("a")], [Symbol("a"), Symbol("b")],
            [Symbol("a"), Symbol("a"), Symbol("b")], [Symbol("a"), Symbol("b"), Symbol("b")],
            [Symbol("c"), Symbol("d")],
            [Symbol("a"), Symbol("b"), Symbol("d")],
            [Symbol("a"), Symbol("b"), Symbol("c"), Symbol("d")],
            [Symbol("a"), Symbol("b"), Symbol("c"), Symbol("c"), Symbol("d")],
            [Symbol("a"), Symbol("b"), Symbol("c"), Symbol("c"), Symbol("c"), Symbol("d")],
            [Symbol("a"), Symbol("b"), Symbol("c"), Symbol("c"), Symbol("c"), Symbol("c"), Symbol("d")]
        ]
        not_accepted = [
            [Symbol("")],
            [Symbol("e")],
            [Symbol("b"), Symbol("a")]
        ]
        assert all(dfa.accepts(word) for word in accepted)
        assert not all(dfa.accepts(word) for word in not_accepted)


class RegexUtilNFATest(unittest.TestCase):
    def setUp(self):
        pass

    def test_check_is_nfa(self):
        two_cycles_graph = build_two_cycle_labeled_graph(first_cycle=4, second_cycle=4, edge_labels=("A", "B"))
        nfa = graph_to_nfa(two_cycles_graph)
        assert not nfa.is_deterministic()
        assert not nfa.is_empty()

    def test_empty_multi_graph_nfa(self):
        empty_graph = MultiGraph()
        nfa = graph_to_nfa(empty_graph)
        assert nfa.is_empty()

    def test_empty_multi_di_graph_nfa(self):
        empty_graph = MultiDiGraph()
        nfa = graph_to_nfa(empty_graph)
        assert nfa.is_empty()

    def test_nfa_equivalent_two_cycles(self):
        expected = NondeterministicFiniteAutomaton()
        start_states = {0, 1, 2}
        final_states = {0, 1, 2}
        for state in start_states:
            expected.add_start_state(State(state))
        for state in final_states:
            expected.add_final_state(State(state))

        expected.add_transitions(
            [
                (0, "A", 1),
                (1, "A", 0),
                (0, "B", 2),
                (2, "B", 0),
            ]
        )
        two_cycles_graph = build_two_cycle_labeled_graph(first_cycle=1, second_cycle=1, edge_labels=("A", "B"))
        actual = graph_to_nfa(two_cycles_graph, start_states, final_states)
        assert actual.is_equivalent_to(expected)
        assert not actual.is_deterministic()
        assert not actual.is_empty()

    # Number of states in actual graph is x2 of that created from Automata...
    @unittest.SkipTest
    def test_load_from_file_two_cycles(self):
        expected = NondeterministicFiniteAutomaton()
        start_states = {0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 60, 61, 62, 63, 64, 65, 66, 67, 68, 69, 70, 71}
        final_states = {0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 60, 61, 62, 63, 64, 65, 66, 67, 68, 69, 70, 71}
        for state in start_states:
            expected.add_start_state(State(state))
        for state in final_states:
            expected.add_final_state(State(state))

        expected.add_transitions(
            [
                (1, "A", 2),
                (2, "A", 3),
                (3, "A", 4),
                (4, "A", 5),
                (5, "A", 6),
                (6, "A", 7),
                (7, "A", 8),
                (8, "A", 9),
                (9, "A", 10),
                (10, "A", 11),
                (11, "A", 12),
                (12, "A", 13),
                (13, "A", 14),
                (14, "A", 15),
                (15, "A", 16),
                (16, "A", 17),
                (17, "A", 18),
                (18, "A", 19),
                (19, "A", 20),
                (20, "A", 21),
                (21, "A", 22),
                (22, "A", 23),
                (23, "A", 24),
                (24, "A", 25),
                (25, "A", 26),
                (26, "A", 27),
                (27, "A", 28),
                (28, "A", 29),
                (29, "A", 30),
                (30, "A", 31),
                (31, "A", 32),
                (32, "A", 33),
                (33, "A", 34),
                (34, "A", 35),
                (35, "A", 36),
                (36, "A", 37),
                (37, "A", 38),
                (38, "A", 39),
                (39, "A", 40),
                (40, "A", 41),
                (41, "A", 42),
                (42, "A", 0),
                (0, "A", 1),
                (0, "B", 43),
                (43, "B", 44),
                (44, "B", 45),
                (45, "B", 46),
                (46, "B", 47),
                (47, "B", 48),
                (48, "B", 49),
                (49, "B", 50),
                (50, "B", 51),
                (51, "B", 52),
                (52, "B", 53),
                (53, "B", 54),
                (54, "B", 55),
                (55, "B", 56),
                (56, "B", 57),
                (57, "B", 58),
                (58, "B", 59),
                (59, "B", 60),
                (60, "B", 61),
                (61, "B", 62),
                (62, "B", 63),
                (63, "B", 64),
                (64, "B", 65),
                (65, "B", 66),
                (66, "B", 67),
                (67, "B", 68),
                (68, "B", 69),
                (69, "B", 70),
                (70, "B", 71),
                (71, "B", 0)
            ]
        )

        two_cycles_graph = read_graph_from_file("expected_2_cycles_graph")
        actual = graph_to_nfa(two_cycles_graph, start_states, final_states)
        assert actual.is_equivalent_to(expected)
        assert not actual.is_deterministic()
        assert not actual.is_empty()

